Revert flash speed changes, leave only xosc
Previous commit changed three board settings in hopes of improving
compatibility with some rev4 hardware. One of them applied to the
crystal oscillator and two to the flash chip. After some more tests in
the field, I now think only the crystal was problematic so this
reverts the flash speed changes.

They didn't necessarily hurt anything and let's keep them in mind in
case we want to use a different flash chip in the future, but for now
let's stick to the defaults until there's a reason to do otherwise.